Updated formatting of tests again, added tests for balance put api
